finished UI

+++ Application/WaffleDB/WaffleDBAPI/TableUtility/ProductAddition.cs
@@ -8,7 +8,7 @@
         private float _price { get; set; }
         public float price
         {
-            get { return (float)(_price * 1.19); }
+            get { return _price * 1.19f; }
 
             set { _price = value; }
